# Replace IsAssignableFrom with is in converters