--- a/src/Microsoft.DotNet.Wpf/src/System.Windows.Controls.Ribbon/Microsoft/Windows/Controls/Ribbon/Primitives/RibbonScrollButtonVisibilityConverter.cs
+++ b/src/Microsoft.DotNet.Wpf/src/System.Windows.Controls.Ribbon/Microsoft/Windows/Controls/Ribbon/Primitives/RibbonScrollButtonVisibilityConverter.cs
@@ -45,25 +45,18 @@ namespace Microsoft.Windows.Controls.Ribbon.Primitives
             //
             // Parameter Validation
             //
-
-            Type doubleType = typeof(double);
             if (parameter == null ||
                 values == null ||
                 values.Length != 4 ||
-                values[0] == null ||
-                values[1] == null ||
-                values[2] == null ||
-                values[3] == null ||
-                !typeof(Visibility).IsAssignableFrom(values[0].GetType()) ||
-                !doubleType.IsAssignableFrom(values[1].GetType()) ||
-                !doubleType.IsAssignableFrom(values[2].GetType()) ||
-                !doubleType.IsAssignableFrom(values[3].GetType()) )
+                values[0] is not Visibility ||
+                values[1] is not double ||
+                values[2] is not double ||
+                values[3] is not double)
             {
                 return DependencyProperty.UnsetValue;
             }
 
-            Type paramType = parameter.GetType();
-            if (!(doubleType.IsAssignableFrom(paramType) || typeof(string).IsAssignableFrom(paramType)))
+            if (parameter is not double && parameter is not string)
             {
                 return DependencyProperty.UnsetValue;
             }