## Description
Replace IsAssignableFrom with is in converters. It is faster, has smaller IL and smaller JIT code.

I used [this benchmark](https://gist.github.com/ThomasGoulet73/86043718584c119dc746b1df108ca871) to ensure that it is in fact faster with `MenuScrollingVisibilityConverter` as an example.

Here is the result:
```
| Method |    values | parameter |     Mean |    Error |   StdDev | Ratio |  Gen 0 | Code Size | Allocated |
|------- |---------- |---------- |---------:|---------:|---------:|------:|-------:|----------:|----------:|
|    Old | Object[4] |         0 | 41.75 ns | 0.623 ns | 0.552 ns |  1.00 | 0.0076 |   1,037 B |      24 B |
|    New | Object[4] |         0 | 14.18 ns | 0.175 ns | 0.146 ns |  0.34 | 0.0076 |     936 B |      24 B |
```

## Customer Impact
None.

## Regression
No.

## Testing
Tested a few converters locally.

## Risk
None.